% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_data.R
\name{get_info_data}
\alias{get_info_data}
\title{Get info  on data}
\usage{
get_info_data(data)
}
\arguments{
\item{data}{a data object}
}
\value{
a lit that contained the dimension and data object variable names
}
\description{
allow to have information on a data set passed in parameter: dimensions and name of columns in a list
}
\examples{
example <- get_info_data(data = iris)
dput(example)

}
