% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_data.R
\name{get_mean_data}
\alias{get_mean_data}
\title{get mean from data}
\usage{
get_mean_data(data)
}
\arguments{
\item{data}{a numeric vector}
}
\value{
mean numeric value for the mean
}
\description{
Get a table with the average of all numeric variables in your dataset as parameter.
(If the input table has no numeric variables, then the function must return an explicit message)
}
\examples{
test <- c(1,2,3)
get_mean_data(data = test)
}
