% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_my_dataset.R
\docType{data}
\name{my_dataset}
\alias{my_dataset}
\title{DATASET_TITLE}
\format{
A data frame with 87 rows and 1 variables:
\describe{
\item{\code{dplyr::pull(dplyr::starwars, species)}}{character COLUMN_DESCRIPTION}
}
}
\usage{
my_dataset
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
